requirements.txt
psycopg2-binary==2.9.9
SQLAlchemy==2.0.27
pandas==2.2.0
python-dotenv==1.0.1
pytest==8.0.0
